requirements.txt
aiohttp>=2.2.3
async-timeout>=1.2.1
asyncpg>=0.11.0
chardet>=3.0.4
discord.py>=1.0.0a1266
multidict>=3.1.0
pytoml>=0.1.14
websockets>=3.3
yarl>=0.11.0
